conform to google via cpplint

+++ utils.cpp
@@ -1,4 +1,5 @@
-#include "main.hpp"
+// NOLINT(legal/copyright)
+#include "./main.hpp"
 
 bool seeded = false;
 std::mt19937 gen;
@@ -13,31 +14,31 @@ size_t sysrandom(void* dst, size_t dstlen) {
 }
 
 void Utils::SeedCSPRNG() {
-	std::uint_least32_t seed;
-	sysrandom(&seed, sizeof(seed));
-	gen = std::mt19937(seed);
-	seeded = true;
+        std::uint_least32_t seed;
+        sysrandom(&seed, sizeof(seed));
+        gen = std::mt19937(seed);
+        seeded = true;
 }
 
 void Utils::ClearBuffer(char buf[]) {
-	// Resets buffer to '\0' until a null byte is encountered.
-	int c = 0;
-	while (buf[c]) buf[c++] = '\0';
+        // Resets buffer to '\0' until a null byte is encountered.
+        int c = 0;
+        while (buf[c]) buf[c++] = '\0';
 }
 
 double Utils::Random() {
-	// Call upon our friend the CSPRNG (cryptographically-secure pseudo-random number generator)
-	// with a lot of code yoinked from StackOverflow.
-	if (!seeded)
-		SeedCSPRNG();
-	return (double) gen() / gen.max();
+        // Call upon our friend the CSPRNG (cryptographically-secure pseudo-random number generator)
+        // with a lot of code yoinked from StackOverflow.
+        if (!seeded)
+                SeedCSPRNG();
+        return static_cast<double>(gen()) / static_cast<double>(gen.max());
 }
 
 ImVec2 Utils::v_abs(ImVec2 __val) {
-	return ImVec2(abs(__val.x), abs(__val.y));
+        return ImVec2(abs(__val.x), abs(__val.y));
 }
 
 bool Utils::Equal(ImVec2 a, ImVec2 b) {
-	return (int) a.x == (int) b.x && (int) a.y == (int) b.y;
+        return a.x == b.x && a.y == b.y;
 }
